rendering on front page, still needs to be fixed
